// designmode-master/src/integrations/extensions/EmbeddedViews/generated-types.d.ts
type InternalPath = string
// ...
type DesignModeEventData = {
  "tool.edit.start": undefined
  "tool.edit.end": undefined
  "selection.changed": {
    paths: InternalPath[]
  }
  "model.changed": {
    rootUrn: Urn
  }
  "colorbar.range.changed": {
    renderScope: string
    rangeFilter: {
      lowerIndex: number
      upperIndex: number
    }
  }
}
// ...
interface ActionAPI {
  add: {
    oneByUrn(
      urn: Urn,
      options?: {
        /** Defaults to current context path */
        parentPath?: InternalPath
        child?: Omit<Child, "urn">
      },
    ): Promise<Action[]>
    multipleByUrn(
      items: {
        urn: Urn
        /** Defaults to current context path */
        parentPath?: InternalPath
        child?: Omit<Child, "urn">
      }[],
    ): Promise<Action[]>
  }
  update: {
    oneByUrn(path: InternalPath, urn: Urn): Promise<Action[]>
  }
  delete: {
    one(path: InternalPath): Action[]
  }
  utils: {
    getPathOfAction: (action: Action) => InternalPath
    generateNewChildKey: () => string
  }
  apply(name: string, actions: Action[]): void
}
interface AnalysisSelectionAPI {
  getTopLevelElementsInsidePolygons: (polygons: [number, number][][]) => string[]
}
interface CameraApi {
  moveCamera: (
    position: {
      x: number
      y: number
      z: number
    },
    target?: {
      x: number
      y: number
      z: number
    },
    zoom?: number,
    transitionTimeMs?: number,
    easingFunction?: (t: number) => number,
  ) => Promise<void>
  switchPerspective(): Promise<void>
  captureScreen: (ctx: CanvasRenderingContext2D, width?: number, height?: number) => Promise<void>
  getCurrentCameraState: () => {
    position: {
      x: number
      y: number
      z: number
    }
    target: {
      x: number
      y: number
      z: number
    }
  }
  getCameraSettings: () => {
    type: "perspective" | "orthographic"
  }
  cameraEvents: {
    addEventListener: (type: CameraEventTypes, callbackFunction: (e: unknown) => void) => void
    removeEventListener: (type: CameraEventTypes, callbackFunction: (e: unknown) => void) => void
  }
}
interface ColorbarAPI {
  add: ({
    colors,
    unit,
    labelPosition,
    labels,
    isInteractive,
  }: {
    colors: string[]
    unit?: string | undefined
    labelPosition?: "center" | "edge" | undefined
    labels?: string[]
    isInteractive: boolean
  }) => void
  remove: () => void
  getRenderScope: () => string
}
interface DesignEventsAPI {
  addListener: <T extends keyof DesignModeEventData>(event: T, callback: EventCallback<T>) => void
  removeListener: <T extends keyof DesignModeEventData>(event: T, callback: EventCallback<T>) => void
}
interface DesignToolAPI {
  getPoint(): Promise<DesignToolVec3 | undefined>
  getPolygon(): Promise<DesignToolVec3[] | undefined>
  getExtrudedPolygon(): Promise<DesignToolExtrudedPolygon | undefined>
  getLine(): Promise<DesignToolLine | undefined>
}
interface ElementColorAPI {
  setColors(colors: Map<string, string>): void
  clearColors(paths: string[]): void
  clearAll(): void
}
interface ElementStateApi {
  proposalId: string
  rootUrn: Urn
  getElement: (urn: Urn) => FormaElement
  getElementByPath: (path: InternalPath) => FormaElement | undefined
  getVolumeMesh: (urn: Urn) => VolumeMesh | undefined
  getFootprint: (urn: Urn) => Footprint | undefined
  getUrnForPath: (path: InternalPath) => Urn | undefined
  getWorldTransform: (path: InternalPath) => {
    elements: Transform
  }
  getGeojson: (urn: Urn) => Feature | undefined
}
interface GroundTextureApi {
  add: (name: string, canvas: HTMLCanvasElement, position: Position, scale?: GroundTextureScale) => void
  updateTextureData: (name: string, canvas: HTMLCanvasElement) => void
  updatePosition: (name: string, position: Position) => void
  remove: (name: string) => void
}
type PathVisibility = {
  path: InternalPath
  visible: boolean
}
interface HiddenElementsAPI {
  setVisibility(path: InternalPath, visible: boolean): void
  setVisibilityBatch(paths: PathVisibility[]): void
  cleanup(): void
}
interface HousingAPI {
  createHousingLine(
    line: HousingLineString,
    buffer?: number,
    placementSide?: "left" | "right",
    templateId?: HousingTemplateId,
  ): Promise<Urn>
  listTemplates(): Promise<HousingTemplate[]>
}
interface PersistenceAPI {
  getIsUrnPersisted(urn: Urn): boolean
}
interface ProjectAPI {
  currentProject?: {
    geoLocation?: [number, number]
    timezone: string
    hubId: string
    id: string
    countryCode: string
    name: string
  }
}
interface ProjectGeoLocationAPI {
  projectGeoLocation?: {
    projString: string
    srid: number
    point: [number, number]
  }
}
interface ProposalAPI {
  changeProposal: (proposalUrn: Urn, revision?: string) => Promise<void>
  changeProposalPageReload: (proposalUrn: Urn, revision?: string) => Promise<void>
  refreshProposal: () => void
}
interface RenderApi {
  upsert: (toAdd: {
    id: string
    geometryData: {
      position?: Float32Array
      uv?: Float32Array
      normal?: Float32Array
      index?: number[]
      color?: Uint8Array
    }
    transform: Transform
    spec: "vertexColorsTransparent"
    mode: "normal"
  }) => void
  remove: (id: string) => void
  cleanup: () => void
}
interface RenderGeojsonApi {
  upsert: (toAdd: { id: string; geojson: FeatureCollection<LineString | Polygon>; transform: Transform }) => void
  remove: (id: string) => void
  cleanup: () => void
}
interface RenderGlbApi {
  upsert: (toAdd: { id: string; glb: ArrayBuffer }) => Promise<void>
  remove: (id: string) => void
  cleanup: () => void
}
interface RenderElementApi {
  upsert: (toAdd: {
    id: string
    rootUrn: Urn
    elements: {
      urn: Urn
      /** Defaults to current context path */
      parentPath?: InternalPath
      transform: Transform
    }[]
  }) => Promise<void>
  remove: (id: string) => void
  cleanup: () => void
}
interface SelectionAPI {
  selection: string[]
  resetSelection: () => void
}
interface SunApi {
  get: () => Date | undefined
  set: (date: Date) => void
}
interface TerrainApi {
  getBbox: Bbox
  elevationAt: (x: number, y: number) => number
}
interface TerrainPadsApi {
  addPads: (operations: TerrainOperation[]) => void
  applyPads: (operations: TerrainOperation[]) => void
  getPads: () => TerrainOperationBase[]
}

declare const context: {
  cameraApi: CameraApi | undefined
  elementStateApi: ElementStateApi | undefined
  projectApi: ProjectAPI | undefined
  proposalApi: ProposalAPI | undefined
  renderApi: RenderApi | undefined
  renderGlbApi: RenderGlbApi | undefined
  renderElementApi: RenderElementApi | undefined
  colorbarApi: ColorbarAPI | undefined
  renderGeojsonApi: RenderGeojsonApi | undefined
  selectionApi: SelectionAPI | undefined
  analysisSelectionApi: AnalysisSelectionAPI | undefined
  sunApi: SunApi | undefined
  terrainApi: TerrainApi | undefined
  terrainPadsApi: TerrainPadsApi | undefined
  actionApi: ActionAPI | undefined
  projectGeoLocationApi: ProjectGeoLocationAPI | undefined
  designToolApi: DesignToolAPI | undefined
  groundTextureApi: GroundTextureApi | undefined
  designEventsApi: DesignEventsAPI | undefined
  elementColorApi: ElementColorAPI | undefined
  hiddenElementsApi: HiddenElementsAPI | undefined
  housingApi: HousingAPI | undefined
  persistenceApi: PersistenceAPI | undefined
}

type EmbeddedViewHostContext = typeof context